pyproject.toml
[project]
name = "hw1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "pandas>=2.2.3",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "setuptools>=75.8.0",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "tqdm>=4.67.1",
    "wandb>=0.19.5",
]
